requirements.txt
opik[proxy]
python-dotenv
pyyaml 
